UI changes (Added RecyclerView for questions list)

+++ app/src/main/java/com/app/basics/daggerhilt/storage/db/Question.kt
@@ -13,8 +13,8 @@ data class Question(
     val viewCount: Int,
     val answerCount: Int,
     val score: Int,
-    val lastActivityDate: Int,
-    val creationDate: Int,
+    val lastActivityDate: Long,
+    val creationDate: Long,
     val questionId: Int,
     val link: String,
     val title: String
